refac(ipipeline): rename the execution package
